* updated issues.md
* all tests pass

- created test/a11yStatusSpec.js (Karma runs separately so won't use Enzyme)

* lint ok
