+++ karma.conf.js
@@ -4,10 +4,10 @@ var webpackConfigBase = require('./webpack.config.base');
 
 module.exports = function(config) {
   config.set({
-    browsers: [process.env.TRAVIS ? 'Chrome_travis_ci' : /*'Chrome',*/ 'Firefox'],
+    browsers: [process.env.TRAVIS ? 'Chrome_travis_ci' : 'Chrome', 'Firefox'],
     client: {
       // Don't show console output.
-      captureConsole: true,
+      captureConsole: false,
     },
     customLaunchers: {
       Chrome_travis_ci: {
@@ -15,17 +15,14 @@ module.exports = function(config) {
         flags: ['--no-sandbox'],
       },
     },
-    files: [
-      'test/components/a11yResultsSpec.js',
-      //'test/index.js'
-    ],
+    files: ['test/a11yStatusSpec.js', 'test/index.js'],
     frameworks: ['mocha', 'chai'],
     preprocessors: {
-      'test/components/a11yResultsSpec.js': ['webpack', 'sourcemap'],
+      'test/a11yStatusSpec.js': ['webpack', 'sourcemap'],
       'test/index.js': ['webpack', 'sourcemap'],
     },
     reporters: ['dots'],
-    singleRun: false,
+    singleRun: true,
     webpack: Object.assign(webpackConfigBase, {
       devtool: 'inline-source-map',
       mode: 'development',
@@ -35,4 +32,3 @@ module.exports = function(config) {
     },
   });
 };
-
