+++ package.json
@@ -9,7 +9,7 @@
     "build": "sh ./scripts/build.sh",
     "check": "npm run lint && npm test",
     "clean": "rm -rf css && rm -rf dist && rm -rf lib",
-    "deploy-example": "bash ./scripts/deployExample.sh",
+    "deploy-example": "sh ./scripts/deployExample.sh",
     "lint": "eslint .",
     "postpublish": "npm run deploy-example",
     "prepublishOnly": "npm run check && npm run build",
